functions/requirements.txt
oauth2client
google-api-python-client
google-cloud-storage